iommu: msm: Remove IRQF_TRIGGER_RISING flag while registering interrupt
While registering global interrupts we were
passing IRQF_TRIGGER_RISING flag to request
interrupt to be rising edge triggered. But it
seems for the particular interrupt line only
level triggered parameters are allowed. This is
because the interrupt distributor supports 1-N
software model for the particular interrupt
whereas we tried to program it to N-N model.

The issue was masked earlier as _setup_irq method
was not returning any error but just masking
the flags to permissible parameters. But with
update in _setup_irq function we are now seeing
error being reported saying failed to set
trigger mode. Fix this by not passing
TRIGGER_RISING flag to request_threaded_irq.

Change-Id: Iecf3f1663d2f8a5a70ddfc9b634398d2845c78e9
Signed-off-by: Susheel Khiani <skhiani@codeaurora.org>