ARM: dts: 15801: Fix overexposure when capturing photos with flash
When photos are captured with flash, there are 2 phases of the snapshot:
pre-flash (low power) and the actual flash (high power). During the low-power
flash phase, 3A acquires data on the lighting and uses it to configure the
actual snapshot with high-power flash. Since the gap between low-power and
high-power flash is so great, 3A overcompensates exposure by assuming the
scene with high-power flash is much darker than it actually is. This results
in the final photo with high-power flash being severely overexposed and
susceptible to motion blur due to the use of a longer-than-necessary
exposure time.

By increasing the current of the LEDs for torch-mode from 48 mA to 100 mA,
the pre-flash phase when capturing a photo appears brighter to 3A, so 3A
won't overestimate the exposure. Photos with flash are now properly exposed.

100 mA was derived empirically using a pitch-black room as the control
variable.

Additionally, the torch is now ~2x brighter as a result.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Scott <scott@ubuntu.com>